requirements.txt
yk_face
Flask
Flask-Cors
python-dotenv

